# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Monday API Automation",
    "description": "API para automatizar a criação de estrutura de pastas baseada em demandas do Monday.com, definindo analistas responsáveis por produto.",
    "version": "1.0.0",
    "contact": {
      "name": "Nathan Silva - Fagron Tech",
      "email": "nathan.silva@fagrontech.com.br"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "https://monday-api-automation.onrender.com",
      "description": "Servidor de Produção (Render) - GitHub: https://github.com/Nathan-Paranhos/monday-api-automation"
    },
    {
      "url": "http://localhost:3000",
      "description": "Servidor de Desenvolvimento Local"
    }
  ],
  "paths": {
...
    "/api/automation/list": {
      "get": {
        "tags": ["Automação"],
        "summary": "Listar Automações",
        "description": "Lista todas as automações executadas",
        "operationId": "listAutomations",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {"type": "integer", "default": 1}
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {"type": "integer", "default": 10}
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de automações",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "automations": {"type": "array"},
                    "total": {"type": "integer"},
                    "page": {"type": "integer"},
                    "limit": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/automation/process": {
      "post": {
        "tags": ["Automação"],
        "summary": "Processar Automação",
        "description": "Executa uma nova automação",
        "operationId": "processAutomation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutomacaoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Automação processada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomacaoSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/produto/{id}": {
      "get": {
        "tags": ["Monday.com"],
        "summary": "Consultar Produto por ID",
        "description": "Busca o produto associado a um cliente específico no Monday.com",
        "operationId": "consultarProduto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do cliente",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 12345
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoResponse"
                },
                "example": {
                  "operacao_status": "ok",
                  "id_cliente": 12345,
                  "produto": "BOT",
                  "principal_produto": "Fórmula Certa",
                  "status": {
                    "original": "Na Fila",
                    "normalizado": "na_fila"
                  },
                  "elemento": "12345 - FARMÁCIA EXEMPLO",
                  "campos": {
                    "data_solicitacao": "2024-01-10",
                    "tipo_rx": "Fórmula Certa"
                  },
                  "responsavel": "Pedro.Ribeiro@fagrontech.com.br,Bruno.Vaz@fagrontech.com.br",
                  "timestamp": "2024-01-15T10:30:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "ID do cliente inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Produto não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/farmacias/bot": {
      "get": {
        "tags": ["Monday.com"],
        "summary": "Buscar Farmácias com Produto BOT",
        "description": "Busca todas as farmácias que possuem o produto 'BOT' no Monday.com",
        "operationId": "buscarFarmaciasBOT",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filtrar por status normalizado",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "produto_principal",
            "in": "query",
            "description": "Filtrar por produto principal",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Farmácias com produto BOT encontradas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmaciasBOTResponse"
                },
                "example": {
                  "status": "ok",
                  "total": 2,
                  "filtros": {
                    "status": "na_fila",
                    "produto_principal": "Fórmula Certa"
                  },
                  "farmacias": [
                    {
                      "id": "123456789",
                      "elemento": "2707 - BOULEVARD PHARMA",
                      "produto": "BOT",
                      "principal_produto": "Fórmula Certa",
                      "status": {
                        "original": "Na Fila",
                        "normalizado": "na_fila"
                      }
                    },
                    {
                      "id": "987654321",
                      "elemento": "3456 - FARMÁCIA CENTRAL",
                      "produto": "BOT",
                      "principal_produto": "Fórmula Certa",
                      "status": {
                        "original": "Na Fila",
                        "normalizado": "na_fila"
                      }
                    }
                  ],
                  "timestamp": "2024-01-15T10:30:00.000Z"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AutomacaoRequest": {
        "type": "object",
        "required": ["id_cliente", "nome_farmacia"],
        "properties": {
          "id_cliente": {
            "type": "integer",
            "minimum": 1,
            "description": "ID único do cliente",
            "example": 12345
          },
          "nome_farmacia": {
            "type": "string",
            "minLength": 1,
            "description": "Nome da farmácia",
            "example": "Farmácia Exemplo"
          }
        }
      },
      "AutomacaoSuccessResponse": {
        "type": "object",
        "properties": {
          "operacao_status": {
            "type": "string",
            "enum": ["ok"],
            "description": "Status da operação"
          },
          "produto": {
            "type": "string",
            "enum": ["BOT", "Fórmula Certa", "Phusion"],
            "description": "Produto encontrado no Monday.com"
          },
          "principal_produto": {
            "type": "string",
            "description": "Principal produto associado ao item no Monday.com"
          },
          "status": {
            "type": "object",
            "properties": {
              "original": {
                "type": "string",
                "description": "Status original do Monday.com"
              },
              "normalizado": {
                "type": "string",
                "description": "Status normalizado para uso interno"
              }
            },
            "description": "Informações de status do item"
          },
          "pasta": {
            "type": "string",
            "description": "Caminho completo da pasta criada"
          },
          "arquivo_modelo": {
            "type": "string",
            "description": "Caminho completo do arquivo modelo copiado"
          },
          "responsavel": {
            "type": "string",
            "description": "Email do responsável pelo produto. Para o produto BOT, contém múltiplos emails separados por vírgula."
          },
          "elemento": {
            "type": "string",
            "description": "Campo elemento do Monday.com que contém o código do cliente e nome da farmácia (formato: 'CÓDIGO - NOME_FARMACIA')",
            "example": "12345 - Farmácia Exemplo"
          },
          "campos": {
            "type": "object",
            "additionalProperties": true,
            "description": "Campos adicionais do Monday.com"
          },
          "cliente": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "description": "ID do cliente"
              },
              "nome_farmacia": {
                "type": "string",
                "description": "Nome da farmácia"
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp da operação"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["erro"],
            "description": "Status da operação"
          },
          "erro": {
            "type": "string",
            "description": "Mensagem de erro descritiva"
          },
          "codigo": {
            "type": "string",
            "enum": [
              "DADOS_INVALIDOS",
              "DEMANDA_NAO_ENCONTRADA",
              "ERRO_CONEXAO",
              "ERRO_PERMISSAO",
              "ERRO_INTERNO",
              "ROTA_NAO_ENCONTRADA",
              "ID_INVALIDO",
              "PRODUTO_NAO_ENCONTRADO"
            ],
            "description": "Código de erro padronizado"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp do erro"
          }
        }
      },
...
      "ProdutoResponse": {
        "type": "object",
        "properties": {
          "operacao_status": {
            "type": "string",
            "enum": ["ok"],
            "description": "Status da consulta"
          },
          "id_cliente": {
            "type": "integer",
            "description": "ID do cliente consultado"
          },
          "produto": {
            "type": "string",
            "enum": ["BOT", "Fórmula Certa", "Phusion"],
            "description": "Produto encontrado"
          },
          "principal_produto": {
            "type": "string",
            "description": "Principal produto associado ao item no Monday.com"
          },
          "status": {
            "type": "object",
            "properties": {
              "original": {
                "type": "string",
                "description": "Status original do Monday.com"
              },
              "normalizado": {
                "type": "string",
                "description": "Status normalizado para uso interno"
              }
            },
            "description": "Informações de status do item"
          },
          "elemento": {
            "type": "string",
            "description": "Campo elemento do Monday.com que contém o código do cliente e nome da farmácia (formato: 'CÓDIGO - NOME_FARMACIA')",
            "example": "12345 - Farmácia Exemplo"
          },
          "campos": {
            "type": "object",
            "additionalProperties": true,
            "description": "Campos adicionais do Monday.com"
          },
          "responsavel": {
            "type": "string",
            "description": "Email do responsável pelo produto. Para o produto BOT, contém múltiplos emails separados por vírgula."
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp da consulta"
          }
        }
      },
      "FarmaciasBOTResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["ok"],
            "description": "Status da operação"
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "description": "Número total de farmácias encontradas com produto BOT"
          },
          "filtros": {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "description": "Filtro de status normalizado aplicado (se houver)",
                "enum": ["na fila", "em andamento", "concluido", "finalizado", "pausado", "cancelado", "relacionamento", "pausado features", "aguardando 3º", "desenvolvimento", "aguardando implantações", "pausado cliente", "inadimplente"]
              },
              "produto_principal": {
                "type": "string",
                "description": "Filtro de produto principal aplicado (se houver)",
                "enum": ["BOT", "Fórmula Certa", "Phusion"]
              }
            },
            "description": "Filtros aplicados na consulta"
          },
          "farmacias": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "ID único do item no Monday.com"
                },
                "elemento": {
                  "type": "string",
                  "description": "Nome da farmácia (ex: 2707 - BOULEVARD PHARMA)"
                },
                "produto": {
                  "type": "string",
                  "enum": ["BOT"],
                  "description": "Produto encontrado (sempre BOT nesta consulta)"
                },
                "principal_produto": {
                  "type": "string",
                  "description": "Principal produto associado ao item no Monday.com"
                },
                "status": {
                  "type": "object",
                  "properties": {
                    "original": {
                      "type": "string",
                      "description": "Status original do Monday.com"
                    },
                    "normalizado": {
                      "type": "string",
                      "description": "Status normalizado para uso interno"
                    }
                  },
                  "description": "Informações de status do item"
                },
                "campos": {
                  "type": "object",
                  "additionalProperties": true,
                  "description": "Campos adicionais do Monday.com"
                }
              },
              "required": ["id", "elemento", "produto"]
            },
            "description": "Lista de farmácias com produto BOT"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp da consulta"
          }
        },
        "required": ["status", "total", "farmacias", "timestamp"]
      },
...
    }
  },
...
}